+++ src/world/generation/perlin.hpp
@@ -6,7 +6,7 @@
 
 class Perlin {
     public:
-        Perlin(uint64_t seed = 32467, float frequency = 0.01, uint16_t octaves = 8, float lacunarity = 2.0, float gain = 0.5);
+        Perlin(int seed = 32467, float frequency = 0.01f, uint16_t octaves = 8, float lacunarity = 2.0, float gain = 0.5);
         void setRange(float low, float high);
         float get(float x, float y);
         float get(float x, float y, float z);
